tests: Move test coverage for app/ into a new executable
